+++ R/namespace.R
@@ -2,5 +2,4 @@
 #' @importFrom coda as.mcmc.list as.mcmc thin mcmc.list
 #' @importFrom graphics plot
 #' @importFrom stats coef terms
-#' @importFrom tibble tibble
 NULL